完善充值

+++ yudao-module-shop/yudao-module-shop-biz/src/main/java/cn/iocoder/yudao/module/shop/controller/admin/recharge/vo/RechargeOrderCreateReqVO.java
@@ -14,14 +14,14 @@ public class RechargeOrderCreateReqVO extends RechargeOrderBaseVO {
 
     @ApiModelProperty(value = "会员编号", required = true)
     @NotNull(message = "会员编号不能为空")
-    private Integer memberId;
+    private Long memberId;
 
     @ApiModelProperty(value = "充值金额", required = true)
     @NotNull(message = "充值金额不能为空")
     private BigDecimal amount;
 
     @ApiModelProperty(value = "充值活动编号")
-    private Integer rechargeId;
+    private Long rechargeId;
 
     @ApiModelProperty(value = "充值活动名称")
     private String rechargeName;
